New feature and small layout changes
New feature: an option checkbox for fast-forwarding indivisible steps. When remainder bits are not divisible by generator bits, XOR calculations can be skipped, because eventually remainder bits will stay the same.
Small style changes to the layout, including a constraint on minimum width of the control panel.
Remove `pages\testform.tsx`. This was used to test and view sample TailwindCSS components.
Small changes to `README.md` to reflect the new feature.
Bump the project version to 1.0.1.